mark clipboard on web marked as wasm32 only and install_text_agent  move to egui_winit
